# optimize LazyBranch range getindex and return VoV